feat: Add configuration file infrastructure
Add infrastructure for configuration of changelog generation.
Configuration can be placed in a json file named "changelog.settings.json" in the repository root directory.
The configuration file is optional. If file does not exist, the default settings are used.

Configuration is implemented using Microsoft.Extensions.Configuration and uses the "layered" approach.
A settings file must not specify all settings but only the settings that are different from the default settings.
At runtime, both the default settings and the settings in the repository are "merged" together.