feat: Add option to specify output path in the configuration file
Add "outputPath" setting that defaults to "changelog.md". When the output path in the configuration file is a relative path, it is interpreted as being relative to the repository directory.

In turn, make the output path command line parameter optional. When output path is specified on command line it overrides the output path in any configuration file.