fix: Ignore duplicate versions
Versions loaded from git tags can be ambiguous because git only ensures the tag name is unique but different semantic version strings might be the same version after normalization. Furthermore, the version parser allows tags to be prefixed and suffixed.

To avoid issues, only the version from the first parsed tag will be used. When a tag is parsed and another parsed tag name yielded the same version, the subsequently processed tag will be ignored. When a duplicate tag is encountered, a warning will be written to the log.