build: Disable packaging for git submodules
Repository contains GitLabApiClient as a submodule because application relies on changes in the library not (yet) available on NuGet. GitLabApiClient in turn contains versioning logic executed during pack which changes the Azure DevOps build number.
To work around this and use the right build number on Azure DevOps, disable NuGet packaging fir GitLabApiClient