fix(gitlab): Fix linking of closed Merge Requests and Milestones
Insertion of links for GitLab Merge Request and Milestone references did not work when the Merge Request or Milestone was closed, because query by default only return open items.

This fix adjusts the queries sent to GitLab to include all Merge Requests and Milestones with a given id regardless of the item's state.