Add .NET 5 build targets
Add a net5.0 build target to the main application but keep the .NET Core 2.1 and .NET Core 3.1 targets as long as these frameworks are supported.
Retarget the "docs" and "schema" helper projects to net5.0 without keeping older targets because these programs are only used within this repository.

Update language version to use C# 9.0