WIP: Allow overriding individual footer display names
Change format of "Footer" setting in the configuration file from an array to an object.
In the configuration object, use the footer name as property name.
The value of the property continues to be a "Footer Configuration" object which was reduced to a single "displayName" property (the name is no longer requried because it is already defined by the property name).