WIP: Allow overriding individual scope display names
Change format of "Scopes" setting in the configuration file from an array to an object.
In the configuration object, use the scope name as property name.
The value of the property continues to be a "Scope Configuration" object which was reduced to a single "displayName" property (the name is no longer required because it is already defined by the property name).