WIP: Allow overriding individual entry type display names
Change format of "Entry Types" setting in the configuration file from an array to an object.
In the configuration object, use the entry type as property name.
The value of the property continues to be a "Entry Type Configuration" object which was reduced to a single "displayName" property (the name is no longer required because it is already defined by the property name).