build: Remove copying of schmea from New-Release script
Do not copy the configuration file schema for the current version when creating a new release.
Because the schema will typcially be referenced though GitHub's raw content url, the version of the schema from the corresponding release branch can be used instead of a copy served from the master branch.