refactor: Use implicit footer to add commit link to entries
Instead of having explicity logic to get a commit's web link in the integration providers and support in the template to include the commit link as "fake" footer, add an implicit footer referencing the commit to each entry.

Adding the link is now handled like any other commit reference in the integration providers and the template uses the same logic for all footers.