test: Fix sporadic test failure in E2ETests
Disable parallelization of all tests that read or modify environment variables and ensure variables are set back to their original value after the test.
Tests in E2ETests were sporadically failing because ChangeLogConfigurationLoader was running in parallel and was modifying the environment which influenced the behavior of the changelog child process launched by E2ETests.

To disable parallelization of these tests, introduce a xunit test collection that all tests that interact with environment variables must be part of. The test collection provides a collection fixture that ensures all relevant variables are cleared before the test and offers helper methods to update and reset environment variables.