Register all template implementations with the DI container and select the right implementation in RenderTemplateTask
Remove the need to conditioally configure the DI container by no longer registering a single instance of ITemplate based on the configuration.
Instead, register all implementations and select the right implementation based in the configuraion in RenderTemplateTask.

This (along with future refactorings) will enable composing the application / creating the DI container independently of the configuration